Clean up Dockerfiles, verify Gitblit download.
Clean up the dockerfiles, changing some unnecessary convoluted things and
bringing them to a modern, more readable format.

Switch to `wget`, which is often used by now and already exists on Alpine.
Use variables to make updating for a new version easier and improve
readability.
Move the whole data directory and place a link to it in the gitblit
directory so that potentially the scripts in the gitblit directory
still work.

The downloaded tarball is checked against its SHA256 has sum to verify
the download. (See official image recommendations:
https://github.com/docker-library/official-images/blob/master/README.md#security)