Create docker controlled settings files.
Set some defaults in properties files that are controlled by docker.
The /opt/gitblit/system.properties file is out of reach for the user.
It can be sued to set defaults specific to the docker image and defaults
for new settings that get introduced in later versions.
The /opt/gitblit/data/gitblit-docker.properties file is in preparation
of being able to control certain settings by environment variables.
I don't believe that the docker image should set defaults that need to
be edited by the user, but instead the common practice of controlling
certain behaviour via env variables should be used. These can then be
reflected by settings in this file, which gets written by some docker
mechanism, e.g. an entry point script.
Of course, Gitblit could itself read from env variables, but that is a
larger change.