Enable shutdown with `gitblit-stop.sh`command.
Clean up by deleting not needed scripts for service start.

For Alpine, change `/bin/bash` to `/bin/sh` so that the scripts can run.

Add `/opt/gitblit` to the path. Together with the `WORKDIR` this enables
calling scripts in `/opt/giblit` in a `docker exec` command.
Alas, not every script does work with Alpine's shell.