Define a Docker volume for Gitblit data
Define a VOLUME for the image path `/var/opt/gitblit`. This provides
an automatic, anonymous Docker volume so that filesystem operations
are more performant than when writing to the RW image layer, and
persists the data on the host between recreating a container.
Gitblit data is moved under that path.

The actual `baseFolder` for Gitblit is `/var/opt/gitblit/etc`. Data is split
into three folders for configuration (`etc/`), repositories (`srv/`)
and temporary file (`temp/`). This way the default is one volume, but more
advanced usages are also possible where multiple different volumes are
used for the different types of data.

Closes issue #9