Store a defaults.properties that is not to be edited
Store a copy of the defaults.properties file under `/opt/gitblit/etc`.
This is included in the settings files chain, instead of the one in
`/var/opt/gitblit/etc` which might get edited by a user. The defaults
file needs to be the one included first. When updating to a new version,
this file will include the latest settings, while the one on the volume
might be at an older state.

An altered copy of the `defaults.properties` file is placed into
`/var/opt/gitblit/etc` for reference. The settings in the system
properties files are deleted or commented out.

Also move the `system.properties` file to `/opt/gitblit/etc`.

The include of the defaults and system properties is moved to the
file `gitblit-docker.properties` to move it out of the way where
the user would set her custom settings in `gitblit.properties`.