Provide env variable GITBLIT_RPC to select level of RPC to enable
Instead of having RPC fully enabled by default, the new default is that
basic RPC is enabled but no management, like in the normal default
configuration of Gitblit.
A new environment variable `GITBLIT_RPC` is introduced that can be
used as a command line argument to control the level of RPC that should
be enabled.

This variable is also provided as a build argument, so that a `-rpc`
variant of the image can be build with `--build-arg GITBLIT_RPC=admin`.