Support mount bound volumes by populating them
For volumes that are provided via a mount bind of a local directory,
the contents are not automatically copied over from the directory
that is getting mounted over. So the entrypoint script will generally
copy all files from a backup data directory into the data directories
that do not yet exist. This way all files should be there but edited
ones not be overwritten.

The defaults.properties file is always copied, since it could contain
new settings upon an upgrade, and it should never be edited by a user.

If running as root, also file ownership is changed to the gitblit
user so that gitblit can read and write data.

Closes #4