Provide a script to migrate configuration data to the new layout.
With the image for version 1.9.0 the directory layout changed, as well
as how some properties are set for defaults and the docker image setup.

If a previous container was run with a volume attached, then the old
layout and data should be migrated to the new layout and container
configuration.

This can be done with the provided `migrate-data` script.