Configure .bashrc with colorised 'll' alias
It is getting annoying when there is no 'll' alias and no colorised
output from ls when opening a bash in the container. When developing,
the occasion that a shell is run in the container are numerous.
So we add the 'll' alias as a default to the root user's .bashrc.