Add a template and script to generate Dockerfiles, also for snapshot images
We only have two Dockerfiles, one for the standard image and one for the
Alpine image, for released Gitblit versions. Now we add a new dimension:
Docker images of snapshot builds, in preparation of providing nightly
builds via a docker image.
Since the variants are mostly the same with only small differences, it
is getting tedious to maintain multiple Dockerfiles in parallel. Instead,
we do what the Big Ones do, just smaller, i.e. generate Dockerfiles from
a common template.

A small, bash-based templating engine is
[Bash-TPL](https://github.com/TekWizely/bash-tpl)
I use this to create a stand-alone bash script, `generate_dockerfiles.sh`
from a template: `Dockerfile.template`. The script should not be edited
manually, but changes should be made to the template and then the
script be regenerated:
`$ bash-tpl -- Dockerfile.template > generate_dockerfiles.sh`
I did not include the `bash-tpl` tool. It can be retrieved from the
Github project if needed.

The `generate_dockerfiles.sh` script is included in the repo since it
is rather static and so that it can be directly used by CI without
first having to generate it. With this script it should be easy enough
to deploy docker images of nightly Gitblit builds in the future.

Github links for associated issues: PR gitblit/gitblit#1232 and
issue gitblit/gitblit#1223