Add Goss test suite for testing the generated docker images
These are for use with Goss @ https://github.com/aelsabbahy/goss,
or rather `dgoss` which is the wrapper for testing Docker containers.
The `goss_wait.yaml` file is used to wait for the Gitblit process to
have started so that testing all ports and HTTP requests can be done.

Thanks go to @nicholaswilde for making Goss known to me.