Use a different path for temp directory, which can be deleted
The containerised version defines a path for the temporary folder to
be used by Gitblit. This is so that one can choose to mount a tmpfs
in the container at this path. But Gitblit tries to delete the folder
when it starts to make sure it is empty. This does not work when the
folder path is a mount point. So instead, a folder `gitblit` inside
the temporary folder is set as temp folder for Gitblit, which can be
deleted and recreated.
This fixes #12