Update links to docker documentation and fix typos
Links have changed in the Docker documentation web pages. While the old
links do redirect, this still updates the links in the Hub readme file
to the new, current pages. Also, because not every redirect was accurate
for the context it was used in our readme.
Also, fix two typos in comments in the entrypoint script, which is why
the goss.yaml file changed, as it tests for the file size of that file.