Transition from Label Schema to OCI labels
The Label Schema defined a common set for labels for docker images.
Now the Open Container Initiative has defined labels under the
org.opencontainers.image namespace, which are to replace the ones from
Label Schema.
https://github.com/opencontainers/image-spec/blob/main/annotations.md

So we transition to the new OCI labels, keeping the old ones around for
a period of time (like a year or so).